Made/moved directories to more closely match scheduler format
