+++ social-cinema/.env.development
@@ -1,0 +1,3 @@
+REACT_APP_WEBSOCKET_URL=ws://localhost:8001
+PORT=8000
+CHOKIDAR_USEPOLLING=false
