+++ social-cinema/src/components/App.css
