+++ social-cinema/src/components/App.js
