+++ social-cinema/src/components/App.test.js
